File renaming and propTypes
