+++ src/index.js
@@ -9,4 +9,4 @@ export {
 
 export { default as Icon } from './Icon'
 export { default as Toggle } from './Toggle'
-export { default as HorizontalSelect } from './Select'
+export { HorizontalSelect } from './Select'
