+++ stories/SwitchStory.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
-import HorizontalSelect from '../src/Select'
+import { HorizontalSelect } from '../src/Select'
 
 import ColorModeComparison from './ColorModeComparison'
